+++ xmake/actions/config/main.lua
@@ -95,6 +95,7 @@ function _need_check(changed)
     if not changed then
         if os.mtime(path.join(os.programdir(), "core", "main.lua")) > os.mtime(config.filepath()) then
             changed = true
+            os.touch(config.filepath(), {mtime = os.time()})
         end
     end
     return changed
@@ -358,7 +359,6 @@ force to build in current directory via run `xmake -P .`]], os.projectdir())
         localcache.clear("option")
         localcache.clear("package")
         localcache.clear("toolchain")
-        localcache.clear("cxxmodules")
         localcache.set("config", "recheck", true)
         localcache.save()