+++ xmake/actions/run/main.lua
@@ -196,7 +196,7 @@ function main()
         -- we need clear the previous config and reload it
         -- to avoid trigger recheck configs
         config.clear()
-        task.run("build", {target = option.get("target"), all = option.get("all")})
+        task.run("build", {target = option.get("target"), all = option.get("all")}, {disable_dump = true})
     end
 
     -- load targets