+++ xmake/rules/c++/xmake.lua
@@ -21,18 +21,8 @@
 rule("c.build")
     set_sourcekinds("cc")
     add_deps("c.build.pcheader", "c.build.optimization", "c.build.sanitizer")
-    on_build_files("private.action.build.object", {batch = true, distcc = true})
+    on_build_files("private.action.build.object", {jobgraph = true, batch = true, distcc = true})
     on_config(function (target)
-        -- enable vs runtime as MD by default
-        if target:is_plat("windows") and not target:get("runtimes") then
-            local vs_runtime_default = target:policy("build.c++.msvc.runtime")
-            if vs_runtime_default and target:has_tool("cc", "cl", "clang_cl") then
-                if is_mode("debug") then
-                    vs_runtime_default = vs_runtime_default .. "d"
-                end
-                target:set("runtimes", vs_runtime_default)
-            end
-        end
         -- https://github.com/xmake-io/xmake/issues/4621
         if target:is_plat("windows") and target:is_static() and target:has_tool("cc", "tcc") then
             target:set("extension", ".a")
@@ -43,21 +33,11 @@ rule("c.build")
 rule("c++.build")
     set_sourcekinds("cxx")
     add_deps("c++.build.pcheader", "c++.build.modules", "c++.build.optimization", "c++.build.sanitizer")
-    on_build_files("private.action.build.object", {batch = true, distcc = true})
+    on_build_files("private.action.build.object", {jobgraph = true, batch = true, distcc = true})
     on_config(function (target)
         -- enable c++ exceptions by default
         if target:is_plat("windows") and not target:get("exceptions") then
             target:set("exceptions", "cxx")
-        end
-        -- enable vs runtime as MD by default
-        if target:is_plat("windows") and not target:get("runtimes") then
-            local vs_runtime_default = target:policy("build.c++.msvc.runtime")
-            if vs_runtime_default and target:has_tool("cxx", "cl", "clang_cl") then
-                if is_mode("debug") then
-                    vs_runtime_default = vs_runtime_default .. "d"
-                end
-                target:set("runtimes", vs_runtime_default)
-            end
         end
         -- https://github.com/xmake-io/xmake/issues/4621
         if target:is_plat("windows") and target:is_static() and target:has_tool("cxx", "tcc") then