+++ xmake/rules/utils/compiler_runtime/xmake.lua
@@ -40,5 +40,16 @@ rule("utils.compiler.runtime")
             end
             target:set("runtimes", runtimes)
         end
+
+        -- enable vs runtime as MD by default
+        if target:is_plat("windows") and not target:get("runtimes") then
+            local vs_runtime_default = target:policy("build.c++.msvc.runtime")
+            if vs_runtime_default and target:has_tool("cxx", "cl", "clang_cl") then
+                if is_mode("debug") then
+                    vs_runtime_default = vs_runtime_default .. "d"
+                end
+                target:set("runtimes", vs_runtime_default)
+            end
+        end
     end)